fetch_and_store_financial_news/requirements.txt
requests
google-cloud-bigquery
vaderSentiment
